web: Fix card prizes
When the 'tuhinakerroin' was utilized in card selection the card prize
was not used as a selection criteria. This led all categories of the
cards (cheap, mid and expensive) to get cards from all random prize
categories.

Fix this by including 'WHERE c.prize <comparison operator> 4' to the
queries. This results the cheap cards to cost less than 4, mid cards 4
and expensive ones more than 4.

Fixes: 0aaa7f0 ("web: Add web interface")
Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>