regulator: bd72720: Use parent dt node
The regulator device-trees are attached to the MFD driver (which is the
real I2C deviece). The reguilator driver is bound to a platform device
which is a child of the MFD driver with the device-tree information.

The regulator driver did try to find some device-tree values from the
platform device's device-tree node, which is missing. Thus the probe
failed.

Use the platform device's parent device's (MFD device's) device tree
node for the regulator properties.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>