power: supply: bd72720: Fix dcin register check
The dcin register's address is stored as 8-bit unsigned value. Value -1
is used to denote that this adress is not valid.

The code which later checks for validity compares the register address
to signed integer -1. This makes the check to effectively be:

255 != -1, which is always true. This results the code to try to use
value 255 as dcin collapse limit register's address. The write fails
because 255 is greater than the max register address for the regmap.

Fix this (quick'n dirty) for now by just checking if

regval != (u8)-1.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>