power: supply: Handle interrupts
Handle the power-supply specific interrupts. The handling is very
limited. Basically the interrupt reasons are printed, and users are
informed that the state has changed. The actual status is then queried
by the users from the registers (outside of the interrupt handlers).

Add the handlers which print reason and ping the power-supply class /
simple gauge.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>