dt-bindings: bd72720: Improve doc
The properties which are expected to be shown in the monitored-battery
node aren't documented in the charger bindings. The charger binding doc
is intended to only document the properties that are in the charger
device tree node. This is a standard way, and mandated by the dt
validation as well.

The BD72720 does however use some non standard bindings from a battery
node. ROHM charger specific bindings don't really belong to the generic
static battery node either. So they aren't documented, which is
"unhelpful".

It still is clear that users do need to have some documentation or our
device/drivers will be unusable. In order to solve this somehow I will
add some doc of battery node to the charder doc as a comment.

NOTE: This does not make dts validation to pass when ROHM specific
properties are present in the battery node. Well, I must just live with
this for now.

Add example of the monitored battery node.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>