regulator: bd72720: Fix regulator data address
The regulator probe loops through the table of regulators. At the
beginning of the loop, the address of current item is stored to a
temporary pointer, which is then used to point to the data of a
regulator which is currently being added.

When the data pointer is stored to regulator frameworks as rdev data,
this pointer is again indexed by the loop index. This doubles the
address offset from the beginning of the array, causing the regulator
data pointer to point wrong indexes, evenetually out of bounds of the
data array.

This leads to runtime issues ranging from undefined behaviour and
crashes to security issues.

Fix the address offset by removing the second indexing of the pointer.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>