web: move style from db.php to header.php
The db.php containing all the page-header stuff is counter intuitive
and confusing. Additionally, updating the db.php is a bit of a pain as
it is necessary to clean the passwords etc. from the db.php while
pushing the updated version to git.

Keeping the mysql connection stuff in db.php file alone makes the need
to update the db.php less frequent.

Create a new 'header.php'-file which contains the header stuff and leave
the db.php to only contain the database stuff.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>