web: index.php: Shorten card table
The card table is lengthy making it hard to see all the picked cards.
Additionally, the prize-type column is confusing for the treasure cards,
because we use same Finnish word 'raha' for the 'treasure'-card type and
the 'money' prize-type (as opposed to loan). Shifting the card-type next
to the card-name, and embedding the prize type in the prize column
clarifies the output a bit.

Move the card-type field next to card name and combine the prize and
prize type fields into one.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>