mfd: bd72720: Enable main level IRQs
The BD72720 has 2 levels of IRQ registers. The first level denotes which
block has active interrupts. The second level specifies an exact IRQ
inside a block.

The BD72720 has enable (mask) registers for both of these levels. The
1.st level mask register was not written at all, causing the interrupts
to stay disabled. The 2.nd level masks are handed by the regmap IRQ.

The easiest way to go is to enable all the 1.st level interrupts at the
probe, and leave them to be enabled. The actual interrupts stay disabled
for as long as the 2.nd level mask is not enabled. This allows us to
leave the whole IRQ masking and acking to be implemented by the standard
Linux regmap-IRQ controller.

Enable all 1.st level IRQs at the MFD core driver probe.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>