mfd: bd72720: Fix IRQ mapping
The mapping from the 1.st level IRQ status bits to 2.nd level IRQ status
registers was wrong. This resulted the IRQs for last IRQ registers to not
work correctly. For example the RTC IRQs were never triggered but the
CPU was hogged to running the ISR with no valid statuses until the
kernel disables the IRQ from the host side.

Fix the IRQ bit to IRQ reg mapping.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>