mva-test: BD72720 test dts, fix runlevels
The runlevel properties in the test dts did not match to what was
implemented in the driver. Change the test device-tree to match the
driver implementation.

This is actually a sub-optimal solution because the set of the
properties that was implemented in the driver:

rohm,dvs-runlevel0-voltage
rohm,dvs-runlevel1-voltage
rohm,dvs-runlevel2-voltage
rohm,dvs-runlevel3-voltage

is more complex and less expansible than what was in the device-tree
(array of voltages). Furthermore, the microvolt suffix could've been
better.

Rationale for taking the driver implementation as it is instead of using
what was in the test device-tree are 100% not technical. This decision
is dictated by the time. It's quicker and less error prone to change the
device-tree than to change the driver.

TODO: Revise this decision if there is extra time.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>