regulator: Fix runlevel control
When the BD72720 regulator driver parses the device tree for the
regulators which should be controlled by the sub run states, it compares
the device tree node name to the match name given in the regulator
descriptor so it can identify which regulator's control is being
configured.

The check however is inverted. Eg, code checks:
if (!of_node_name_eq(np, rd[i].desc.of_match))
when it should check:
if (of_node_name_eq(np, rd[i].desc.of_match))

This makes the code to assume given properties are for wrong regulators.
In the end this can cause invalid memory accesses as the driver assumes
regulator run level control specific entries to be set also on such
regulators which don't support this functionality.

Fix the node name check.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>