regulator: bd72720: Fix sub run level control
There were a few problems in the sub run level control code.

First of all, the regmap is obtained from the driver data for the
runlevel control. For other uses it is typically obtained from the rdev.
The regmap pointer was not initialized in driver data, leading to
crashes.

Secondly, all the runlevel control variables were written to buck1 data,
whether the source of runlevel control information was the LDO1 or buck1
node.

Fix these issues by:

1. initializing the regmap pointer in driver data during the probe.
2. use correct data pointer for runlevel control information when the
   device-tree is parsed for the runlevel properties.

Additionally, add some debug prints to runlevel control functions.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>