mfd: BD72720: Drop of_compatible from leds
There is no real reason to have a compatible property in the leds node.
The BD72720 does include the LED hardware block, and it is really enough
to tell the PMIC is BD72720.

From the Linux software side this means that the platform driver
handling the LEDs will just fetch the device-tree from the parent MFD
node. This keeps the behaviour same as regulators have, making the
software more consistent.

Having the compatible in existing device trees (for example for the
BD71828) does not really hurt though. Hence the binding document is kept
intact.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>