gpio: bd72720: Support EPDEN and static mapping
The BD72720 driver did not support EPDEN pin (GPIO) configuration. The
EPDEN pin configuration register is identical to the GPIO1...5
configuration, so handling the EPDEN can be done in same way as handling
the other GPOs.

Add EPDEN handling to the GPIO driver.

Additionally, the BD72720 GPIO driver did dynamically allocate number of
GPIOs based on the device-tree information during driver probe. This
makes mapping the physical pins to the GPIO numbers hard, because the
meapping will now depend on the device-tree (which should reflect the
used OTP configuration).

Simple solution is to use static mapping by always telling there is 6
GPIOs (GPIO1..5 + EPDEN), and setting the GPIO 'valid mask' based on the
device-tree. This way users can rely on the GPIO numbering to not depend
on OTP. The GPIO1 will be BD72720 GPIO base, GPIO2 will be this base +
1, and finally, the EPDEN will be the base + 5.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>