iio: adc: ti-ads7924: Respect device tree config
The ti-ads7924 driver ignores the device-tree ADC channel specification
and always exposes all 4 channels to users whether they are present in
the device-tree or not. Additionally, the "reg" values in the channel
nodes are ignored, although an error is printed if they are out of range.
The probe is failed only if all of the found channels contained bad
or no "reg" property.

Register only the channels described in the device-tree, and use the reg
property as a channel ID. Fail the probe if bad "reg" properties are
found.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---

This patch shouldn't be sent upstream unless asked for. If you see this,
then I've messed up my patches... Sorry.

Revision history:
v4 => v5:
 - drop dev parameter from ads7924_get_channels_config() and
 - drop num_channels initialization as suggested by David
v3 => v4:
 - Adapt to 'drop diff-channel support' changes to ADC-helpers
 - select ADC helpers in the Kconfig
v2 => v3: New patch

Please note that this is potentially breaking existing users if they
have wrong values in the device-tree. I believe the device-tree should
ideally be respected, and if it says device X has only one channel, then
we should believe it and not register 4. Well, we don't live in the
ideal world, so even though I believe this is TheRightThingToDo - it may
cause havoc because correct device-tree has not been required from the
day 1. So, please review and test and apply at your own risk :)

As a side note, this might warrant a fixes tag but the adc-helper -stuff
is hardly worth to be backported... (And I've already exceeded my time
budget with this series - hence I'll leave crafting backportable fix to
TI people ;) )

This has only been compile tested! All testing is highly appreciated.