mfd: bd72720: Support LID event from HALL IRQ
The BD72720 has a pin which is dedicated for a HALL sensor that is meant
to generate an IRQ when LID is opened/closed.

Add handling of the HALL IRQ to BD72720 so LID event is sent when the
HALL sensor detects a change.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
Please note that this commit is untested as I don't have hardware with
the HALL sensor. This commit is compile tested only.