misc: bd71828-lid: Support BD72720.
The bd71828-lid driver reads the HALL pin state from the PMIC's
registers. The BD72720 has different register address for the HALL
status compared to the BD71828.

Support reading different registers on different PMICs by storing the
device type (PMIC ID) in the  platform ID table, and selecting the status
register based on the PMIC ID.

Store the corret register address in the driver's private data and use
it when deciding the state of the LID.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>