power: supply: bd71827: Fix uninitialized pointer in OCV getter
When battery information is given via module parameters, the
bd71827_get_ocv() function does not initialize driver data (pwr) pointer
before using it. Fix this by moving the check for NULL sw pointer and
the initialization of the pwr pointer out of the

if (!use_load_bat_params || !ocv_table[0])
{ ... }

 - block.

NOTE: Using the module parameters is not suggested method as adding the
new module parameters to upstream kernel is not well received. Any
upstreaming effort for this driver would require the module parameter
code to be dropped. Thus the module parameter code for this driver is not
thoroughly, tested and is likely to be dropped from future versions
supporting new ICs. Upstream practices seem to suggest delivering the
battery details via device-tree, or, by coding the tables in driver and
selecting correct tables when the device is known.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: b7e63cbc4535 ("power: supply: Add bd718(15/27/28/78) charger driver")

---
100% Untested