power: supply: bd71827: Avoid warning about unused dcin_collapse
Setting the dcin_collapse limit is currently not supported by the
BD71827 driver. The range for possible dcin_collapse limits is still
defined in the driver, causing a warning about the unused structure.
Keeping the dcin_collapse limit definition in the driver can be helpful
for the customer, should they need to implement support for setting the
limit. Hence, removing the limit structure from this driver version
might not be the best fix, even though carrying unsupported code is
typically not the right thing to do.

Remove the warning using GCC 'unused' attribute.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>