power: supply: simple-gauge: staticize SIMPLE_GAUGE_PROPS and mark unused
The SIMPLE_GAUGE_PROPS can be made static to limit the scope of the
symbol. It then also needs to be marked as potentially unused because
not all of the users of the simple_gauge.h header refer it - which would
generate a warning if SIMPLE_GAUGE_PROPS was not marked unused.

Make SIMPLE_GAUGE_PROPS static and use GCC attribute ((unused)) to mark
the SIMPLE_GAUGE_PROPS as potentially unused.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>