iio: adc: ad7476: Simplify scale handling
The ad7476 driver supports variants with different amount of supply
regulators. On some variants there is only VCC, which is used as a
reference voltage. Others have separate VREF regulator, and some rely on
internal VREF. Some have both internal VREF and option to connect
external one.

The ad7476 driver reads the regulator voltage only when the user asks to
get the scale. This means the driver needs to do some dancing while
picking the correct reference regulator (or internal reference), and
store it for the later use.

According to the discussion:
https://lore.kernel.org/linux-iio/20250331122247.05c6b09d@jic23-huawei/
variable voltage references are rare, making it hard to justify the
added complexity for supporting those.

Drop the support for the variable voltage references and simplify things
by using the managed regulator get and enable interfaces.

NOTE: This approach may alter the order of powering-up the supplies. All
'non reference' supplies are powered-on before the reference supply.
Some devices require the I/O voltage to be smaller (or at least not much
higher than) the VCC. I don't think the driver currently supports
devices which use VCC as a reference, while also having a separate
VDRIVE.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
I am unsure if the note describing ordering of the supply power-on makes
this change equal to asking for a nose-bleed... Please, advice :)

Furthermore, requiring the VREF to be explicitly specified (instead of
building the heuristics based on the has_vref flag and presence of
internal reference voltage) increases the lines - although, in my biased
opinion, eases the understanding.

NOTE II: This is compile-tested only. Please, don't apply as such. I will
do _some_ testing with ROHM BD79105 if this idea seems to be Ok regardless
the things mentioned above, and send non RFC version.

Still, all testing and logic reviewing is greatly appreciated as always.